boboml-main/requirements.txt
asgiref==3.6.0
contourpy==1.0.6
cycler==0.11.0
Django==3.2.16
djangorestframework==3.14.0
fonttools==4.38.0
joblib==1.2.0
kiwisolver==1.4.4
numpy==1.21.6
packaging==23.0
pandas==1.3.5
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.3
threadpoolctl==3.1.0
waitress==2.1.2
